Add scheduler
