+++ .github/workflows/ci.yml
@@ -115,53 +115,10 @@ jobs:
       - name: Build release
         run: make release
   test:
-    name: Test
-    runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        tz: ["UTC", "Asia/Shanghai", "America/Los_Angeles"]
-    steps:
-      - name: Set timezone
-        run: sudo timedatectl set-timezone ${{ matrix.tz }}
-      - name: Check out code into the Go module directory
-        uses: actions/checkout@v3
-      - name: Cache Go Modules
-        uses: actions/cache@v3
-        id: cache-go
-        with:
-          path: |
-            ~/.cache/go-build
-            ~/go/pkg/mod
-          key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
-          restore-keys: |
-            ${{ runner.os }}-go-
-      - name: Cache tools
-        uses: actions/cache@v3
-        id: cache-tool
-        with:
-          path: bin
-          key: ${{ runner.os }}-test-tool-${{ hashFiles('**version.mk') }}
-          restore-keys: |
-            ${{ runner.os }}-test-tool-
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 16.15
-          cache: 'npm'
-          cache-dependency-path: ui/package-lock.json
-      - name: Install Go
-        uses: actions/setup-go@v3
-        with:
-          go-version-file: 'go.mod'
-      - name: Update dependencies 
-        if: steps.cache-go.outputs.cache-hit != 'true'
-        run: GOPROXY=https://proxy.golang.org go mod download
-      - name: Generate mocks
-        run: make generate
-      - name: Test
-        run: |
-          TEST_CI_OPTS="--cover --covermode atomic --coverprofile=coverage.out --label-filter !slow" make test-ci
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v3
+    uses: ./.github/workflows/test.yml
+    with:
+      options: --label-filter \\!slow
+      
   result:
     name: Continuous Integration
     runs-on: ubuntu-20.04
