+++ banyand/tsdb/metric.go
@@ -23,6 +23,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promauto"
 
 	"github.com/apache/skywalking-banyandb/banyand/observability"
+	"github.com/apache/skywalking-banyandb/pkg/logger"
 )
 
 const statInterval = 5 * time.Second
@@ -50,23 +51,7 @@ func init() {
 	)
 }
 
-func (s *shard) runStat() {
-	go func() {
-		defer s.closer.Done()
-		ticker := time.NewTicker(statInterval)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ticker.C:
-				s.stat()
-			case <-s.closer.CloseNotify():
-				return
-			}
-		}
-	}()
-}
-
-func (s *shard) stat() {
+func (s *shard) stat(_ time.Time, _ *logger.Logger) bool {
 	defer func() {
 		if r := recover(); r != nil {
 			s.l.Warn().Interface("r", r).Msg("recovered")
@@ -101,6 +86,7 @@ func (s *shard) stat() {
 		s.curry(mtBytes).WithLabelValues(name).Set(float64(bs.MemBytes))
 		s.curry(maxMtBytes).WithLabelValues(name).Set(float64(bs.MaxMemBytes))
 	}
+	return true
 }
 
 func (s *shard) curry(gv *prometheus.GaugeVec) *prometheus.GaugeVec {
